requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-rest-swagger
djoser==2.1.0
drf-yasg==1.20.0
flake8==3.9.1
idna==2.10
isort==5.8.0
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
uritemplate
urllib3==1.26.4
